Complete rewrite to support Launchpad X
Not much of the original library remains, as I only used it to test if
it was compatible and just slowly re-wrote the whole thing. The user
interface is now much nicer, and it uses a state-machine grid system
that asynchronously "draws" the desired state onto the pad. This uses
the DAW I/O channel, and not the MIDI channel, meaning events (pad
presses) are uninterrupted.